A3_RL/requirements.txt
tqdm

numpy
matplotlib

pinocchio
adam-robotics[casadi]==0.3.0
l4casadi

torch